Notification_shell_scripting
Shell scripts to handle postfix installation and server start up, fully interactive. If the user already uses a postfix SMTP server, this script will keep a copy of the original /etc/postfix/main.cf file and then temporarily modify it for the duration of the server on time. Once the server is turned off, the main.cf will be reverted. If the server is abruptly turned off, the server will know and handle the situation upon start up.